Change ChangelogGeneratorTest so it passes
Change from using StdOut.capturedLines to capturedString.
The lines captured with the emoji results in weird things, for
example the number "4" preceding the checkmark emoji and
the "Wrote changelog" text. By changing to capturedString,
the wonkyness seems to go away, since the test now passes
when including the check emoji.